+++ Events/CarrierJumpedEvent.cs
@@ -113,10 +113,11 @@ namespace EddiEvents
         public List<EconomyShare> carrierEconomies { get; private set; }
 
         public CarrierJumpedEvent(DateTime timestamp, string systemName, ulong systemAddress, decimal x, decimal y, decimal z,
-            string bodyName, long? bodyId, BodyType bodyType, Faction systemFaction, List<Faction> factions, List<Conflict> conflicts,
-            Economy systemEconomy, Economy systemEconomy2, SecurityLevel systemSecurity, long? systemPopulation, Power powerplayPower,
-            PowerplayState powerplayState, bool docked, string carrierName, StationModel carrierType, long? carrierId, Faction stationFaction,
-            List<StationService> stationServices, List<EconomyShare> stationEconomies) : base(timestamp, NAME)
+            string bodyName, long? bodyId, BodyType bodyType, bool docked, 
+            string carrierName, StationModel carrierType, long? carrierId, List<StationService> stationServices,
+            Faction systemFaction, Faction stationFaction, List<Faction> factions, List<Conflict> conflicts,
+            List<EconomyShare> stationEconomies, Economy systemEconomy, Economy systemEconomy2, SecurityLevel systemSecurity, long? systemPopulation, 
+            Power powerplayPower, PowerplayState powerplayState, bool taxi, bool multicrew, bool inSRV, bool onFoot) : base(timestamp, NAME)
         {
             // System
             this.systemname = systemName;